refactor: utils 내 localStorage관련 파일 이름 수정
